03-Voice-Native-Real-Time-Agent/voice_agent.py
import os
from dotenv import load_dotenv
# Note: Real-time voice often requires specific libraries like PyAudio 
# and cloud SDKs (OpenAI, Deepgram, ElevenLabs).

load_dotenv()

class VoiceAgent:
    def __init__(self):
        self.api_key = os.getenv("OPENAI_API_KEY")
        print("Voice Agent Initialized.")

    def start_listening(self):
        """Placeholder for microphone input logic"""
        print("Listening for audio...")

    def process_and_respond(self, audio_chunk):
        """Placeholder for STT -> LLM -> TTS pipeline"""
        # 1. Transcribe (Whisper)
        # 2. Generate Response (GPT-4o)
        # 3. Synthesize (ElevenLabs)
        pass

if __name__ == "__main__":
    agent = VoiceAgent()
    print("Voice-Native Agent boilerplate ready.")
    print("To implement fully, integrate with OpenAI Realtime API or LiveKit.")
