03-Voice-Native-Real-Time-Agent/requirements.txt
python-dotenv
openai
requests
# Additional libs like pyaudio or livekit may be needed for hardware access
